// Front-end/iglu/src/app/api/models/precioComparado.model.ts
export interface PrecioComparado {
  imagen_producto: string;
  imagen: string;
  nom_producto: string;
  razon_social: string;
  precio: number;
  estado_precio: string;
  precio_mas_barato: number;
  estado_super: string;
  supermercado_mas_conveniente: string;
  monto_total_supermercado: number;
}